requirements.txt
python-dotenv
pandas==0.25.3
mlflow==1.4.0
scikit-learn==0.21.3
google-cloud-storage
virtualenv